removed dead code, and renamed and moved to correct directories files
