+++ assets/js/components/ActivityPageComponent.js
