--- assets/js/components/ActivityPageLayout.js
@@ -1,15 +1,0 @@
-/**
- * Created by ingrskar on 3/6/2017.
- */
-import React from 'react';
-import ActivityPageContainer from '../container/ActivityPageContainer';
-
-
-const ActivityPageLayout = () => (
-    <div>
-        <ActivityPageContainer />
-    </div>
-
-);
-
-export default ActivityPageLayout;
